added entity saving to H2 DB

+++ src/main/java/com/webCFD/webCFDapp/repository/HeatGenerationFieldRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.webCFD.webCFDapp.entities.HeatGenerationFieldEntity;
 
-public interface HeatGenerationFieldRepository extends JpaRepository<HeatGenerationFieldEntity, Double> {
+public interface HeatGenerationFieldRepository extends JpaRepository<HeatGenerationFieldEntity, Long> {
 
 }
